# api/openapi.yaml
openapi: 3.1.0
info:
  title: App Parceiros API
  version: 0.2.0
  description: >
    Contrato da API do App Parceiros. É a fonte da verdade: altere aqui
    antes de implementar e gere os clientes a partir deste arquivo.


    Erros vêm sempre como `{code, message}` (schema `Error`): `code` é
    estável e serve para o programa; `message` já vem em português, pronta
    para mostrar ao usuário. O status segue o tipo do erro: 400 para JSON
    ou parâmetro malformado (`invalid_json`, `invalid_request`), 401 sem
    login, 402 workspace suspenso, 403 sem permissão, 404 não encontrado,
    409 conflito, 410 link vencido, 422 dado inválido, 429 limite de
    chamadas, 502 provedor externo fora do ar, 503 recurso não configurado
    no servidor e 500 `internal_error`.


    As rotas de um workspace ficam sob `/v1/workspaces/{workspaceId}`. Quem
    não é membro recebe 404 `workspace_not_found`. Com o workspace suspenso
    por falta de pagamento, só continuam valendo ver o workspace, sair dele
    (ou remover afiliados) e a assinatura; o resto responde 402
    (`workspace_suspended`, ou `workspace_suspended_owner` para o dono).
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/me/shopee:
    get:
      summary: Situação da conexão com a Shopee
      description: >
        A credencial (AppID e Secret da Open API de afiliados) pertence ao
        usuário, não ao workspace. O Secret nunca aparece em resposta.
      operationId: getShopeeConnection
      tags: [shopee]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Situação
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopeeConnection" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
    put:
      summary: Conectar a conta de afiliado
      description: >
        Valida o AppID e o Secret com uma chamada de teste à Shopee e guarda a
        credencial cifrada (envelope encryption). Substitui a credencial
        anterior, se houver.
      operationId: connectShopee
      tags: [shopee]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConnectShopee" }
      responses:
        "200":
          description: Conectado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ShopeeConnection" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422":
          description: >
            `invalid_app_id`, `invalid_secret`, `invalid_shopee_credential`
            (a Shopee recusou o AppID ou o Secret) ou `shopee_access_denied`
            (a conta não tem acesso à Open API)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: "`shopee_rate_limited` (a Shopee limitou as chamadas; tente de novo em instantes)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "502":
          description: "`shopee_unavailable`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      summary: Desconectar (apaga a credencial)
      operationId: disconnectShopee
      tags: [shopee]
      security: [{ bearer: [] }]
      responses:
        "204": { description: Desconectado }
        "401": { $ref: "#/components/responses/Unauthenticated" }

  # --- radar ---------------------------------------------------------------
...
  /v1/me/results/sync:
    get:
      summary: Situação da sincronização de conversões do usuário
      operationId: getResultsSync
      tags: [results]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Situação
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversionSync" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
    post:
      summary: Sincronizar as conversões agora
      description: >
        Enfileira a leitura do conversionReport da Shopee (últimos 89 dias) com
        a credencial do usuário. A sincronização também roda uma vez por dia.
        Se uma já está em andamento, devolve a situação dela.
      operationId: requestResultsSync
      tags: [results]
      security: [{ bearer: [] }]
      responses:
        "202":
          description: Sincronização pedida
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConversionSync" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "409":
          description: "`sync_no_credential`: a Shopee não está conectada"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: "`synced_recently`: a última terminou há menos de 10 minutos"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Token de acesso do provedor de identidade: JWT do Zitadel (OIDC),
        token opaco do provedor interno (`/v1/auth/login`) ou `dev:<sub>` no
        modo de desenvolvimento.
...
  responses:
    Unauthenticated:
      description: Sem token ou token inválido (`unauthenticated`)
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidJSON:
      description: Corpo que não é JSON, com campo desconhecido ou grande demais (`invalid_json`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            Código estável para o front. Genéricos: unauthenticated,
            internal_error, invalid_json, invalid_request. Auth (provedor
            interno): invalid_credentials, email_taken, invalid_email,
            invalid_name, weak_password, invalid_auth_token,
            too_many_attempts. Contas: workspace_not_found, member_not_found,
            invite_not_found, forbidden, mentorship_only,
            consent_mentorship_only, owner_cannot_leave, no_seats,
            seats_in_use, seats_above_plan, workspace_suspended,
            workspace_suspended_owner, already_member, invite_expired,
            invite_used, invite_revoked, invite_other_email,
            invalid_workspace_name, invalid_photo_url, invalid_invite_email,
            invalid_invite_validity. Assinatura: billing_owner_only,
            billing_managers_only, already_subscribed, no_subscription,
            invalid_tax_id, minimum_seats, no_price, simulation_unavailable,
            billing_unavailable, invalid_billing_webhook. Shopee:
            invalid_shopee_credential, shopee_access_denied,
            shopee_rate_limited, shopee_unavailable, invalid_app_id,
            invalid_secret. Catálogo e radar: product_not_found,
            invalid_page, invalid_per_page, negative_price,
            invalid_min_commission, invalid_min_rating, invalid_sort,
            query_too_long, invalid_history_days, invalid_product_link,
            short_link. Coleções: item_not_found, collection_not_found,
            collection_exists, no_credential, import_unavailable,
            link_rate_limited, product_lookup_failed,
            product_or_link_required, invalid_item_title,
            invalid_item_description, invalid_item_notes, invalid_tag,
            too_many_tags, invalid_item_status, invalid_affiliate_link,
            invalid_collection_name, too_many_item_collections,
            invalid_item_page. Notificações: notification_not_found,
            push_unavailable, invalid_push_subscription,
            too_many_push_subscriptions, invalid_notification_preferences.
            Mídia: video_not_found, invalid_video_link, video_unavailable,
            platform_unavailable, usage_rights_required,
            invalid_video_format, invalid_video_size, video_too_large,
            video_quota_exceeded, uploads_unavailable, upload_closed,
            upload_incomplete, upload_parts_required, invalid_part_number,
            video_not_ready, video_owner_only, video_share_managers_only,
            video_title_too_long, embed_not_downloadable. Curadoria:
            list_not_found, list_item_not_found, video_not_in_list,
            list_managers_only, lists_mentorship_only, list_limit,
            list_empty, already_in_list, list_full, list_not_published,
            invalid_list_title, invalid_list_description,
            invalid_list_comment, list_comment_required,
            invalid_list_order, product_not_in_list. Resultados:
            group_results_forbidden, group_results_mentorship_only,
            sync_no_credential, synced_recently, invalid_period,
            sync_credential_refused, sync_shopee_unavailable.
        message: { type: string, description: Texto em português pronto para mostrar ao usuário }
...
    ConnectShopee:
      type: object
      required: [app_id, secret]
      additionalProperties: false
      properties:
        app_id: { type: string, pattern: "^[0-9]{1,20}$" }
        secret: { type: string, minLength: 1, maxLength: 256, writeOnly: true }
    ShopeeConnection:
      type: object
      required: [status, app_id, verified_at]
      properties:
        status:
          type: string
          enum: [disconnected, connected, invalid, expired]
          description: >
            `invalid`: a Shopee recusou a assinatura (Secret trocado).
            `expired`: a Shopee negou acesso à conta (bloqueada, sem permissão).
        app_id: { type: [string, "null"], description: "AppID mascarado, ex.: ••••5678" }
        verified_at: { type: [string, "null"], format: date-time }

    # radar
...
    ConversionSync:
      type: object
      required: [status, requested_at, finished_at, conversions, error_code, error]
      properties:
        status: { type: string, enum: [never, syncing, ok, error, no_credential] }
        requested_at: { type: [string, "null"], format: date-time }
        finished_at:
          type: [string, "null"]
          format: date-time
          description: Fim da última sincronização que deu certo.
        conversions: { type: integer, description: Itens de pedido lidos na última sincronização que deu certo. }
        error_code:
          type: [string, "null"]
          description: "Código do erro da última, quando falhou (ex.: sync_credential_refused, sync_shopee_unavailable)."
        error: { type: [string, "null"], description: Texto em português de `error_code`. }
...